% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{calmr_ui}
\alias{calmr_ui}
\title{Build calmr.app UI}
\usage{
calmr_ui(analytics_file = "inst/app_resources/google_analytics.html")
}
\arguments{
\item{analytics_file}{A path to google analytics.
Default is NULL, in which case, no analytics are pushed.}
}
\description{
Build calmr.app UI
}
