% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{df_to_parlist}
\alias{df_to_parlist}
\title{Convert parameter data.frame to list}
\usage{
df_to_parlist(df, type)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{type}{A character specifying which type of
list we're dealing with.}
}
\description{
Convert parameter data.frame to list
}
