% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Run the calm shiny application}
\usage{
launch_app(browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{browser}{If TRUE, the app is launched in a browser.}
}
\description{
Run the calm shiny application
}
