% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{make_par_tables}
\alias{make_par_tables}
\title{Parse parameter list into stimulus-specific and global parameters}
\usage{
make_par_tables(model, parameters)
}
\arguments{
\item{model}{A modelname string}

\item{parameters}{A list with parameters}
}
\value{
A list with stimulus and global parameters (both data.frames)
}
\description{
Parse parameter list into stimulus-specific and global parameters
}
