% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{make_timing_tables}
\alias{make_timing_tables}
\title{Parse timing parameter list into data.frames}
\usage{
make_timing_tables(timings)
}
\arguments{
\item{parameters}{A list with timings, as returned by \code{calmr::get_timings()}}
}
\value{
A list with data.frames
}
\description{
Parse timing parameter list into data.frames
}
